+++ CHANGES.md
@@ -5,6 +5,7 @@
 * Add plugins for LESS and SASS sources.
 * Fix source maps of stylesheets and templates compiled with esbuild.
 * Improve performance of counting lines in source stylesheets and templates.
+* Improve performance of minifying by using just the best minifier.
 
 ## 0.4.0